Some code cleanup done using the IDE, fixes #6271

+++ engine/src/main/java/org/apache/hop/pipeline/ModPartitioner.java
@@ -94,7 +94,7 @@ public class ModPartitioner extends BasePartitioner {
         if (longValue == null) {
           value = valueMeta.hashCode(valueData);
         } else {
-          value = longValue.longValue();
+          value = longValue;
         }
         break;
       default:
